GS-50 refactored into component
